# PERF: improve anynan reduction performance by up to 4x on float32